Upgrade third_party/acorn to 8.0.4
This allows us to remove both numeric-separator and
logical-assignment plugin extensions, which reached stage 4
and are supported since Acorn 7.4.0:
https://github.com/acornjs/acorn/blob/master/acorn/CHANGELOG.md#740-2020-08-03
The parser then requires the ECMA_VERSION to be set to 2021 to make it
parse these new constructs.

Next to that, the only breaking change in 8.0.0
(https://github.com/acornjs/acorn/blob/master/acorn/CHANGELOG.md#800-2020-08-12)
makes ecmaVersion required, which we always specified already.

R=bmeurer@chromium.org

Change-Id: I64f06817f2c02e79fc29bd0a49b57b9172977a93