Fix keyboard usage of docking menu
This problem was caused by the TS-ificiation and restructuring of the
code which lead to events being too eagerly consumed by the soft context
menu. In the event that part of the menu is a custom element, we need to
do nothing and _not_ consume the event, so that the custom element gets
the event to consume and process.

Fixed: 1156022
Change-Id: I9889a295d5e7181f53dd9a09ea2fc360f5daae29