[Wasm debug] Autostep over function prologues
When pausing execution, the debugger might stop in a location that's
inside the wasm function prologue. At that point, variables are not
fully initialized, producing a weird UX. This CL uses new information on
the getFunctionInfo plugin API to detect that the debugger is paused in
a bad location and auto-step out of it.

Bug: chromium:1155185
Change-Id: Icb6ea3ff48ff6d8bfbdab2385c4bbe13a9ea5cdb