Fix network category filter being reset if text filter is persisted
This fixes a bug where if the user filters the network panel for the
text "foo", and filters by resource type (e.g. to "XHR"), when they
close and re-open DevTools the search text "foo" is persisted and added
back, but as a side-effect the resource type filter is lost.

I discovered this is true for all the additional filters; the moment you
restore a text filter every other setting is lost. To fix this we avoid
calling the `setTextFilterValue` function, which resets everything else,
and instead set the value of the filter input directly, with no other
side-effects.

Fixed: 1155564
Change-Id: I9fa1db634e9e5ec71e84b220c66503a4ab3e4be5