[Flex] Adding gap support to the flex overlay
Flexbox now also supports *gap properties like css grid.

When gap is added in the main direction (using gap, column-gap in
horizontal writing-mode, or row-gap in vertical writing-mode), then
space is added in between items (on a flex line), where justify-content
space normally exists. Both justify-content and gap spaces work together
to space items apart.

When gap is added in the cross direction (using gap, row-gap in
horizontal writing-mode or column-gap in vertical writing-mode), then
space is added in between flex lines, where align-content space normally
exists. Both align-content and gap spaces work together to space lines
apart.

In the tools, we have 2 main use cases:
- Either draw all empty space at once, with the same style. This happens
with the inspect element mode for example. In this case, we draw the
entire empty area at once, to make the pattern look nice.
- Or draw either of the 4 types of empty spaces. This happens on hover
in the Styles pane, when hovering over align-content, justify-content or
a gap property. In this case, we want to draw just 1 type of space
without the rest.

The first use case is simple as we can draw in one go and just skip the
areas where flex items are.

The second use case made the code a little more complicated.
If drawing the justify-content space, then we need to draw the space on
each flex line, skipping where items are, and also skipping where gaps
are, if any.
Similarly, for align-content, we need to draw the space between flex
lines, but skip where gaps are, if any.

Screenshot of the 4 types of spaces: https://imgur.com/zMVTHyI.png

Bug: 1150832
Change-Id: Iefc9dfd037d12e10d4f3905e54da6d105593ef34