[debugger] Log timings of language extension plugin methods.
This is a (temporary) measure to gain a better understanding of the
timing of the various language plugin methods, and how these affect the
overall experience. By adding this to the frontend itself (the log
messages will only show up in DevTools on DevTools), we also gain the
ability to ask partners / other developers to look at them in case of
performance issues.

Bug: chromium:1083146
Change-Id: I444c48f20be72c0f2b57f7dc821a772e97c5e86f