New: image preview should also show file size
Increase the file size when previewing images in the Elements panel and
the Styles panel. It is consistent with the file size in the Network panel.

This way, we don’t need to check the file size of the images separately in the Network panel, which helps us check whether the images is compressed.

And improved the following user experience:

- The different sizes are displayed in separate rows.
- Click the images to open it in a new tab.
- The currentSrc is renamed Current source, and the width will not exceed 20em to avoid the dialog being too wide. Click to open in a new tab.

Screenshot: https://i.imgur.com/jdGiDXi.png

The next plan is to support the background image to display the width and height of the container.

Bug: 1149832
Change-Id: I3b4fe452938688e61c28769afab836f9a9d71db3