Fix that cross-origin images cannot be saved
Currently, downloading images in imageView used `<a download>`, but Chrome 65[1] has begun to disable the download of cross-origin images.

Data URLs and Blob are not subject to this restriction, so this CL is changed to Data URL to save images. In the future, when the File System Access API matures, we can consider switching over.

Bug: 1157399

Change-Id: I7e4929d0d531e029328ccd70ec91ef6d228b09cd