[Loc] Fail presubmit on modifications to en-US.json
Currently presubmit doesn't fail when it modifies en-US.json. This
allows for the possibility of uploading patches with en-US.json out of
sync with strings found in the product.

This patch creates a new presubmit check that encapsulates checks
whose results can be affected by the output of other checks.

Bug: 1136655
Change-Id: Ia6bddd69a5c323ce81b24c308f0aabf222771e02