Refactor ApplicationPanelTreeElement
Currently we have a large number of classes in ApplicationPanelSidebar
that all inherit (ultimately) from BaseStorageTreeElement. That class-
name is from a type when the Application panel was only about storage
and misleading. This CL renames the class to ApplicationPanelTreeElement
and moves it to its own file. This enables moving other classes to
separate files as well.

Bug: chromium:1157484
Change-Id: Ibe4542774b15daee2e2f78d3499208d8d0c6d7c4