Handle Trust Token done event arriving before requestWillBeSent
Should a Trust Token operation fail or succeed before we fire the
`requestWillBeSent` event, then the Trust Token done event is lost
as the corresponding NetworkRequest instance has not yet been
created.

This CL introduces a new Map, that acts as a cache for such instances.
The Map records early events and `requestWillBeSent` considers
the Map when building the NetworkRequest.

R=sigurds@chromium.org

Bug: chromium:1126824
Change-Id: I7e6b6e7cdd0d6cc893b2ab45e6ef0948d0d55bbb