Refactor ExpandableApplicationPanelTreeElement
Some classes in ApplicationPanelSidebar inherit from a class called
StorageCategoryTreeElement. That class name is from a time when the
Application panel was only about storage and is nowadays misleading.
This CL renames the class to ExpandableApplicationPanelTreeElement
and moves it to ApplicationPanelTreeElement.ts. This enables
moving other classes to separate files later on.

R=wolfi@chromium.org

Bug: chromium:1157484
Change-Id: I3121827da92830ceb1b5132489fd5b0ecedeaf63