Report Trust Token operation result in network panel
This CL adds the 'Results' portion of the Trust Token tab in the
Network panel.

Please note that the issuer and top-level origin are reported by
the result event due to the way the backend implementation. The two
origin are considered part of the parameters though, and are
displayed in the 'Parameters' section.

Screenshots: https://imgur.com/a/XceM0Xl

R=petermueller@chromium.org

Bug: chromium:1126824
Change-Id: I23e6a5dcddb4be8ffbe95dc25bee75ffaedc2064