[module.json extensions] Add 'localizableTitle' field to ViewRegistration interface
Views that contain a defined location can be shown using the Command
Menu. The command added for this purpose follows the shape 'Show 
<ViewTitle>'. Because these commands are localized, a localizable 
version of each view's title should also be provided. 

This CL adds the field 'localizableTitle' for this purpose while keeping
the original 'title' field with the localized string, since it is still
needed to provide the string value to the GRDP system.

Finally, commands added to show migrated views are updated to follow the
same shape as they do for the legacy ones.

Bug: 1134103
Change-Id: I47aedf2ca4c35ed38c0b9be14969946e582afc7d