Don't display RowBucket when cursor is over squiggly line
In the Source Tab you can get the errors associated with a line by hovering over the error icon or the squiggly line.
However the later is really hard to hover since it is very thin.
This CL removes this behaviour and now only the error icon shows the information.

Bug: chromium:1150883
Change-Id: I8cede9987861ace9f48bac4532b880170be73a8d