New: Undefined CSS var() is set to gray
The title is no longer displayed as Jump to definition. I don’t think it makes much sense. The underline usually means that it can be clicked. There is no need to display the text on this key title. It is more practical to display the actual value of the custom attribute.

Bug: 1147750
Change-Id: I33d8b37de2dc6911580f979e31ee61ebcd518e8f