New: undefined CSS var() is set to gray
The title is no longer displayed as Jump to definition. I don’t think it makes much sense. The underline usually means that it can be clicked.
There is no need to display these words in this critical position. It is more practical to display the actual value of the custom attribute.

Screenshot: https://i.imgur.com/BjVx3So.png

Bug: 1124707
Change-Id: I33d8b37de2dc6911580f979e31ee61ebcd518e8f