Network: Fix Har Importer
This CL fixes the requestsFromHARLog function. When the HARImporter.js
file was Typesciptified, it more strictly type checked access to
the maps generated when parsing the HAR files. However, this caused
the function to skip some intended fallbacks that were allowed by
the poorly typed JS: missing pagerefs in the HAR cause skips all together
instead of falling back to using the entry.request.url

This was causing the network tool to be unable to import HAR files it created.

This change restores the fallback behavior while maintainging typechecking.
(Repro steps in bug)

Bug: 1158442
Change-Id: I7bc52fb90b91c4288dea1dd00702797a2ef4d0bd