Use a ResourceTreeFrame's security origin as context in console
Previously, for generated OOPIF we selected the first available url
of the Runtime.CallFrame as a context in the console for context
selection. This change uses the security origin of the
ResourceTreeFrame instead. In the future we are going to
remove the url that is sent along the Runtime.CallFrame.
This fix allows us to have similar behavior to before, even
if the url is no longer available.

Bug: chromium:1159332, chromium:1132260
Change-Id: I54d690075f6767f54da60c1dc203d00e6760b83f