[Lighthouse] Handle element screenshot in node linkifying.
LH now prepends a screenshot to every node detail, cropped to the
element under scrutiny. In CDT, we augment these node details to
link to the node in the Element panel. The implementation clears out
the html to do this. We must now save some of it (the screenshot element)
so we can add it back.

https://imgur.com/yjOaxXc

Change-Id: I88be490709c05ce497296b26683fc3251d066fa8