Removed code that caused text to overlap in 'Throttling' section of DevTools Settings.
In this patch I removed an old flex box property in inspectorCommon.css that
caused the 'optional' text to overlap. Previously when a user navigated to chrome://settings/fonts and set the Minimum font size to 24 the 'optional' text  in the 'Throttling' section of DevTools Settings would overlap. In the throttlingSettingsTab.css file I changed the position property in the conditions-edit-optional class. Text no longer overlaps in 'Throttling' section of DevTools Settings.

R=tvanderlippe@chromium.org

Bug: 1152980
Change-Id: I0e0b992c88bee0a2b69c15cb669c28bc4dc43b71