Convert blocked time to milliseconds when exporting request to HAR
When building the timing information for HAR exports, the
NetworkRequest's startTime, issueTime and endTime are given in seconds,
whereas the values in the NetworkRequests's ResourceTiming property are
given in milliseconds. For HAR exports the former need to be converted
into ms because that is the unit HAR files require.

For the code path when there is no response received, this conversion to
ms has been missing previously.

Fixed: 1145177

Change-Id: I541d6dc636debdc84194381cf34fe8197b0d1d7e