Emulate DevTools environment in component docs
This CL takes the code from the test `EnvironmentHelpers` and moves it
into `front_end/mock_environment`. This is so we can depend on this code
from the component docs folder as well as in test, and can emulate a
DevTools environment in the component docs.

Most of the time we won't need this fully fledged environment, but for
some things (such as context menus, guess what I'm working on right now
:D) we do need a faked out environment to enable these features to run
when we run the component in isolation in the component docs.

Note: while this CL contains no component docs changes that take
advantage of it, I've tested locally with context menus in the data grid
and this change does work.

Fixed: 1148323
Change-Id: I8559a2926cc6ab2bfeba757317bfa019f13f273d