[FrameTree] Persist selection of top frame across reloads
When checking which frame tree item to select, the item's URLs are
compared. This change sets the 'lastSelectedItemPath' for top frames to
'frame://' only. This ensures that when the last selected item is the
top frame, closing and re-opening DevTools on a different page will
still result in the top frame being automatically selected.

Fixed: 1159682

Change-Id: I8eddcd3c0ad8e2e52ef272dc7d176d47186beaa1