[module.json extensions] Add check for duplicate actions and views
To prevent duplicates amongst legacy and pre-registered views and
actions a check is implemented in the place where both legacy and new
extensions of registered in the same data structure and an error is
thrown in the case of duplicates. Note that this check had been
implemented already for setting extensions.

This is only to keep support for both the legacy and new extension
registration systems during the migration. Because a check for
duplicates already exists for extensions registered with the new system,
checks added in this CL will be deleted once all extensions have been
migrated and the legacy system has been completely removed.

Bug: 1134103
Change-Id: I3601dffe64cd2ccfb7c16e3ad10bdc0cb103b189