[Lighthouse] Avoid reprocessing the same Network events.
Also adds test `lighthouse-block-url.js`, which passes
even without this change. I wanted confidence that any
future refactor here will retain the benefits of multi-client.

In short: DevTools is sending the "same" messages to Lighthouse
for the root session: one without a session id (good, expected, LH and
DT assumes this means it comes from the main target), and one also
for the main target but with an explicit sessionId (bad, we end
up double-processing things). This is an overlooked side effect of the
"multi client" work:

- https://chromium-review.googlesource.com/c/chromium/src/+/1743411
- https://chromium-review.googlesource.com/c/chromium/src/+/1838310

This CL is a rather narrow change, but it's important as we are
currently double-counting the size of most requests, which
is a big problem for accuracy. There are likely issues with other domains
sending multiple events. A more involved fix will require a closer
look at the protocol connection juggling going on here.

See https://github.com/GoogleChrome/lighthouse/issues/11415

Change-Id: Ife7296dbd08153784ef700e761d529ffe4b39272