Set expander icon size in pixels
The Layout pane uses the <details><summary> elements to display
expandable sections.
These contain a triangle icon that can be clicked to expand/collapse.
The icon is coming by default with the <summary> element, and we only
style it to look like the other expander icons throughout devtools.

It used to have a font-size in rem applied to it to look correct.
The problem is that this made it look different on different computers
that may have changed the default font-size in chrome://settings/fonts.
Setting the font size in px fixes that.

Before: https://imgur.com/wfIufF4
After: https://imgur.com/A9TimVU

Bug: 1161662
Change-Id: I7496e4e2d7c65a74490954125f857521dc52d1e4