Fix changes computation for inspector-stylesheet
The inspector-stylesheet is a special virtual file that is used
to keep track of any new CSS rules, which can be added by clicking
the "New Style Rule" button in the Elements panel. However, the
changes computation broke and was no longer showing any changes
to this file.

The most likely culprit is https://crrev.com/c/2378070 which changed
the way errors were handled. The TLDR is that the `baseline` became
an empty string, which is falsy. Hence, the diff computation bailed
and was not showing any diffs.

Instead, we should be explicitly checking of the `baseline` is `null`,
since it is legal to perform a diff computation between an empty
string and a larger string.

R=szuend@chromium.org

Fixed: 1152507
Change-Id: I1aa778e7922586a4381cbc1e1e6e543a5ccffdd9