Service Worker Update Cycle Timeline
This CL adds SW update cycle timeline to the Service Workers
view, Application pane. See the explainers below for
requirements.

A summary of features implemented:
- For the service worker associated with the current target, for
every version still alive (including redundant ones), show a
timeline of the install, wait, or activate events.
- The timeline is a waterfall style, different than in the explainer
but similar to that in network timing.
- Such a timeline is based on the observation in the frontend and
is intended to show for example, if an event takes extra long
time in waiting. It is not meant to be a measurement of execution
time of the event handler at the backend.
- Each event's timeline may be expanded to reveal details of
the event handler, such as the starting and ending time of
the event handling.
- Other details in the explainer are not included.


crbug.com/1066604

Explainer:
https://docs.google.com/document/d/1M2xKrlFHrojGHIJ9AVpQdGrmUUU0G8YVToa3kZz3RZA/edit?usp=sharing

Change-Id: I0b3725b342eeb32470ef759bab8a7da5a350dba9