Add body context menu to data-grid
This CL adds support for the context menu that appears when you right
click on the body row of the data-grid.

This menu shows columns you can sort by and you can toggle column
visibility via "Header Options" (the text is identical to the legacy
data grid). It also provides a hook for the user to customise and add
their own items to the context menu.

Bug: 1150793
Change-Id: I6a6d8c6c67156a575477c1e6d438261b85d9b034