Fix sub menu context menu interaction tests on Mac
The interaction tests that looked for submenu entry points in a context
menu would fail on Mac, because on Mac they have the ▶ symbol, whereas
on other platforms they do not.

This CL provides a helper to search for the item with and without the ▶
and ensures the tests pass on all platforms.

Change-Id: I619ff370fbc6388ec166d9fd4b7e944788dbbf6d