Styles: Prevent invisible properties to invalidate visible ones
e.g.
if your css looks like this:
.selector {
  -webkit-background-clip: text;
}

The backend will return two properties '-webkit-background-clip' and
'background-clip', but 'background-clip' is not visible in the styles
panel.

But even though 'background-clip' is not visible, it invalidates
'-webkit-background-clip'.

This behavior is not only for '-webkit-background-clip' there are
more '-webkit-' properties with the same behavior.

This CL fix this issue.

Before: https://i.imgur.com/0MfEApU.jpg
After: https://i.imgur.com/Mc1Wu72.jpg

Bug: 1163314
Change-Id: I1e35bd6cf2df297823ce313cc21d466825201add