This PR will prevent not visible properties to invalidate visible ones.
e.g.
if your css looks like this:
.selector {
  -webkit-background-clip: text;
}

The backend will return two properties '-webkit-background-clip' and
'background-clip', but 'background-clip' is not visible in the styles
panel.

But even though 'background-clip' is not visible, it is
invalidating (only visually crossing out the property, the property value
is propertly applied) '-webkit-background-clip'

This behavior is not only for '-webkit-background-clip' there are
more '-webkit-' properties with the same behavior.

This CL fix this issue.

Bug: 1163314
Change-Id: I1e35bd6cf2df297823ce313cc21d466825201add