Limit Datagrid Screen Reader Alerts to Focused Grids
This CL addresses an issue where datagrids could unexpectedly announce
content to screen readers if a grid node is selected when the grid
does not have focus - which can occur when a grid is first created.

This change limits datagrid alerts to when the grid has focus.

Repro steps in bug description.

Bug: 1164034
Change-Id: Ib6fd7012782f7073617ed20d5432333d83732ccb