[Dark mode]: generate dark styles for cmdevtools.css
Uses the new script to generate the dark mode styles for cmdevtools so save us
manually processing it. Tested by clicking around the sources panel in light and
dark mode and verifying that things look as expected.

Also fixed some bugs in the generator script to ensure we generate the
right selectors.

Bug: 1152736
Change-Id: If3491651be116033be7248e695ceab1de5f3fc5c