Show SameParty cookie attribute in cookies table
The SameParty attribute is a boolean cookie attribute used to indicate
whether a cookie should be included in requests to origins of the
same "first party set". This CL adds the attribute to the CookieTable
similar to how the "Secure" attribute is shown. Editing is not
yet supported.

Screenshots: https://imgur.com/a/GoZj7UQ

R=sigurds@chromium.org

Bug: chromium:1161427
Change-Id: Ic208e8f1411fe1de755c42607190d187302ec21c