Device emulation: re-apply changes after edit
When the UA list changes, re-apply the current device, in case it's
what got edited --- previously the changes would not apply unless one
switch away from the device and back to it again.

This also adds an integration test for the editing functionality,
including that it applies without any fiddling of switching devices
back and forth.

Bug: 1158834
Change-Id: I6bd83b26a00e2fc742fbef6b30c26e06c8ff9c20