Perf panel: fix request finish time
When providing the details for a request in the performance panel, we
calculate the network duration based on the request's finish time minus
its start time. However for data URLs there is no request finish time
since these never go out to the network, which results in a negative
time shown as the code defaults the finish time to zero (0).

This CL updates the code to default the finish time to be the start
time in the event that no finish time exists or if it is zero(0).

Bug: 1157396
Change-Id: I04b03fdeb7941009e88707df44ec68ef9e2f0fda