Use custom CSS property rather than string interpolation
Within LitHtml components we should try to avoid interpolation inside
`style` tags and instead use custom properties that we can set with
JavaScript.

This CL moves one example of this over as a proof of concept, and I will
follow it up with a CL that enforces this pattern by linting against
interpolation, and updating the component docs we have accordingly.

Bug: chromium:1166108
Change-Id: I304df148194b750e1377785a78855a070587a38e