[FrameTree] Show document resource for JS iframes
Previously when writing to an iframe via JS's document.write(), the
update to the frame tree in the Application panel would be incomplete.
It would not show the new document resource for the frame being written
to. This is fixed by this CL.

Deactivate test CL: https://crrev.com/c/2600982
Re-activate test CL: https://crrev.com/c/2600740

Bug: 1107766

Change-Id: I564c50cf5fba7fe327814963f59a74f1bc9a938e