[FrameTree] Show resources for JS iframes
This CL takes care of rendering the resources of an iframe which has
been written to via JS's 'document.write()'. It finds the relevant frame
and then adds this frame and its resources recursively to the
ResourceTreeModel.

The document for the frame being written to is set to the document of
the frame containing the 'document.write()' call. Because showing this
document would be confusing, this is handled by showing 'Document not
available' instead.

Bug: 1107766

Change-Id: I564c50cf5fba7fe327814963f59a74f1bc9a938e