Fixing ModuleUIStrings fallback
Fallback to ModuleUIStrings is returning the ID instead of the string due
to a comparing the en-US string (as declared on the code) with the
localized version when trying to do the reverse lookup.

Fix requires to load the en-US resources to correctly do the reverse
lookup and then search on the right localization file using the id.

Change-Id: Ie66b5afd6a1a0c8998d664fcd17ab73f06be10b9