Fix styling of Developer Resources
The developer resources tab was converted to a meta-file, but it turns
out that its styling got broken: crbug.com/1167625

The root cause was that the `data_grid` CSS files were not loaded yet
and thus were causing issues when `developer_resources-meta.js` would
have loaded before that had happened.

In the long term, we are going to figure out a solution to properly
introduce CSS dependencies in our JS dependency graph, but for now
this should be the lowest risk change to fix the beta branch.

R=jacktfranklin@chromium.org

Fixed: 1167625
Change-Id: I2aea8e0c81c0a92eb30f8a963517d5bf4411ed37