Fix the size of the select in Layout pane
Initially, we had multiple select boxes in the Layout panel and they
were limited to the min width of 150px using the grid layout. Now we
don't need the grid layout we can use max width of 150px and min width of 0
to make sure that the select element shrinks properly.

Fixed: 1160629
Change-Id: Ia583000e9c97de22d809d73b6427671897135767