Localize settings strings when used
Settings strings registered via metafiles were localized when loading the
settings to the application, meanwhile the i18n system needs the
settings to be loaded before a locale is and, thus, before the i18n API
is ready for invocation. This introduces a deadlock in which both
systems require the other to be loaded in order to load.

This CL changes the approach so that settings strings are localized when
used and not when the settings are loaded.

Fixed: 1168257
Change-Id: I62955d270de68ca5d5afb962516dd5131f6ad91a